requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
decorator==5.1.1
Django==4.2.4
django-ckeditor==6.7.0
django-cron==0.6.0
django-js-asset==2.1.0
exceptiongroup==1.1.3
executing==1.2.0
frozenlist==1.4.0
geckodriver-autoinstaller==0.1.0
gTTS==2.3.2
h11==0.14.0
idna==3.4
ipdb==0.13.13
ipython==8.14.0
jedi==0.19.0
matplotlib-inline==0.1.6
multidict==6.0.4
openai==0.27.8
outcome==1.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PySocks==1.7.1
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tqdm==4.66.1
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
undetected-chromedriver==3.5.2
urllib3==2.0.4
wcwidth==0.2.6
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
